% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocomplete.R
\name{autocomplete}
\alias{autocomplete}
\title{Get autocompletion suggestions}
\usage{
autocomplete(text, latitude = NULL, longitude = NULL,
  locale = Sys.getenv("YELP_LOCALE", "en_US"),
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{text}{A string to find completions for.}

\item{latitude}{A number representing the latitude to search close to.
Required if you want business results, but can be \code{NULL} otherwise.}

\item{longitude}{A number representing the longitude to search close to.
Required if you want business results, but can be \code{NULL} otherwise.}

\item{locale}{A string naming the locale. See \code{\link{SUPPORTED_LOCALES}}
for allowed values.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\value{
A list of three elements.
\describe{
\item{terms}{A character vector of completion of the text itself}
\item{businesses}{A data frame of business name matches.}
\item{categories}{A data frame of category matches.}
}
}
\description{
Get term, business, and category autocompletion suggestions.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
(sleep_in_seattle <- autocomplete(
  "sleep", latitude = 47.6, longitude = -122.33
))
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/autocomplete}
}
