% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business-lookup.R
\name{business_lookup}
\alias{business_lookup}
\title{Look up a business by Yelp ID}
\usage{
business_lookup(id, locale = "en_US",
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{id}{A string describing the Yelp ID of a business, as returned by
\code{\link{business_search}}.}
}
\value{
A data frame of detailed business information. In addition to the
fields provided by the business search, you also get opening hours,
URLs of up to 3 photos, whether or not the business has been claimed by
its owner, and whether or not the business has been permanently closed.
}
\description{
Look up a business by Yelp ID
}
\examples{
business_lookup("jDPKMwOtvPA0lkqRwRvSJQ")
}
