% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business-match.R
\name{business_match}
\alias{business_match}
\title{Match a business}
\usage{
business_match(name, city, state, country, address1 = NULL,
  address2 = NULL, address3 = NULL, latitude = NULL,
  longitude = NULL, phone = NULL, zip_code = NULL,
  yelp_business_id = NULL, match_threshold = c("default", "none"),
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{name}{A string giving the name of the business.}

\item{city}{A string giving the city where the business is located.}

\item{state}{A string giving the ISO 3166-2 state code where the business is
located. See \code{\link{ISO_3166_2_CODES}}.}

\item{country}{A string giving the ISO 3166-2 country code where the business
is located. See \code{\link{ISO_3166_2_CODES}}.}

\item{address1}{A string giving the address where the business is located.
Optional, but strongly recommended, especially when using
\code{match_threshold = "default"}.}

\item{address2}{A string giving the address where the business is located.}

\item{address3}{A string giving the address where the business is located.}

\item{latitude}{A number representing the latitude to search close to.}

\item{longitude}{A number representing the longitude to search close to.}

\item{phone}{A string representing a phone number. This should include the
country code and consist of a plus sign followed by only digits, for example
"+1234567890".}

\item{zip_code}{A string giving the zip or postal code where the business is
located.}

\item{yelp_business_id}{A string describing the Yelp ID of a business, as returned by
\code{\link{business_search}}.}

\item{match_threshold}{By default, matches are filtered rather strictly.
If you don't get any results, try setting this to \code{"none"}.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\description{
Find a business, when you have precise info like name and address.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
walmart <- business_match("Walmart", "Albany", "NY", "US", "Washington Ave")
if(interactive()) View(walmart) else str(walmart)
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/business_match}
}
