% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business-search.R
\name{business_search}
\alias{business_search}
\title{Search for businesses}
\usage{
business_search(term, location, latitude = NULL, longitude = NULL,
  radius_m = 40000, categories = NULL, locale = "en_US", limit = 20,
  offset = 0, sort_by = c("best_match", "rating", "review_count",
  "distance"), price = 1:4, open_now = FALSE, open_at = NULL,
  attributes = NULL, access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{term}{A string denoting the search term.}

\item{location}{A string describing the location. If this is not provided,
then \code{latitude} and \code{longitude} are compulsory.}

\item{latitude}{A number representing the latitude to search close to.}

\item{longitude}{A number representing the longitude to search close to.}

\item{radius_m}{A number giving the radius, in metres, of the search circle
around the specified location.}

\item{categories}{A character vector of search categories to filter on,
or \code{NULL} to return everything. See \code{\link{SUPPORTED_CATEGORY_ALIASES}}
for allowed values.}

\item{locale}{A string naming the locale. See \code{\link{SUPPORTED_LOCALES}}
for allowed values.}

\item{limit}{An integer giving the maximum number of businesses to return.
Maximum 50.}

\item{offset}{An integer giving the number of businesses to skip before
returning. Allows you to return more than 50 businesses (split between
multiple searches).}

\item{sort_by}{A string naming the metric to order results by.}

\item{price}{A vector of integers in 1 (cheap) to 4 (expensive) denoting
price brackets.}

\item{open_now}{A logical value of whether or not to only return businesses
that are currently open.}

\item{open_at}{A time when to check if businesses are open.}

\item{attributes}{A character vector of business attributes to filter on.
See \code{\link{SUPPORTED_BUSINESS_ATTRIBUTES}}.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\value{
A data frame of business results.
}
\description{
Use the Yelp business search API to find businesses close to a give location.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
delis_in_queens <- business_search("deli", "Queens, New York")
if(interactive()) View(delis_in_queens) else str(delis_in_queens)
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/business_search}
}
