% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{categories}
\alias{categories}
\title{Describe category alias}
\usage{
categories(alias, locale = Sys.getenv("YELP_LOCALE", "en_US"),
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{alias}{An alias for a search category, chosen from
\code{SUPPORTED_CATEGORY_ALIASES}.}

\item{locale}{A string naming the locale. See \code{\link{SUPPORTED_LOCALES}}
for allowed values.}
}
\value{
A data frame with 1 row and 5 columns.
}
\description{
Provides detailed information about a category alias.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
categories("barcrawl")
categories("drug")
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/category}
}
