% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-lookup.R
\name{event_lookup}
\alias{event_lookup}
\title{Look up an event by Yelp ID}
\usage{
event_lookup(yelp_event_id, locale = get_yelp_locale(),
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{yelp_event_id}{A string denoting a Yelp event ID, as returned by
\code{\link{event_search}}.}

\item{locale}{A string naming the locale. See \code{\link{SUPPORTED_LOCALES}}
for allowed values.}

\item{A}{data frame with 1 row and 26 columns.}
}
\description{
Use the ID from a business search to return only details for that business.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
# First search for events
events_in_kansas_city <- event_search("kansas city")
# Use the returned ID to lookup a specific event
the_main_event <- event_lookup(events_in_kansas_city$id[1L])
if(interactive()) View(the_main_event) else str(the_main_event)
}
}
