% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-search.R
\name{event_to_df_row}
\alias{event_to_df_row}
\title{importFrom magrittr %$%
importFrom tibble data_frame}
\usage{
event_to_df_row(event)
}
\description{
importFrom magrittr %$%
importFrom tibble data_frame
}
