% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured-events.R
\name{featured_event}
\alias{featured_event}
\title{Find featured events}
\usage{
featured_event(location = NULL, latitude = NULL, longitude = NULL,
  locale = Sys.getenv("YELP_LOCALE", "en_US"),
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{location}{A string describing the location.}

\item{latitude}{A number representing the latitude to search close to.}

\item{longitude}{A number representing the longitude to search close to.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\value{
A data frame with 26 columns. Each row corresponds to one event.
}
\description{
List the featured events for a location, as curated by the Yelp staff.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
the_main_event <- featured_event("san francisco")
if(interactive()) View(the_main_event) else str(the_main_event)
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/event_search}
}
