% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured-events.R
\name{featured_events}
\alias{featured_events}
\title{Find featured events}
\usage{
featured_events(location = NULL, latitude = NULL, longitude = NULL,
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{location}{A string describing the location.}

\item{latitude}{A number representing the latitude to search close to.}

\item{longitude}{A number representing the longitude to search close to.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\description{
List the featured events for a location, as curated by the Yelp staff.
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/event_search}
}
