% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-delivery-search.R
\name{food_delivery_search}
\alias{food_delivery_search}
\title{Find food delivery services}
\usage{
food_delivery_search(location, latitude = NULL, longitude = NULL,
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{location}{A string describing the building or other precise location.
If this is not provided, then \code{latitude} and \code{longitude} are compulsory.}

\item{latitude}{A number representing the latitude to search close to.}

\item{longitude}{A number representing the longitude to search close to.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\description{
Find services that deliver food to a specified location in the USA.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
lunchtime <- food_delivery_search("empire state building")
if(interactive()) View(lunchtime) else str(lunchtime)
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/transaction_search}
}
