% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_id_migration}
\alias{handle_id_migration}
\title{Deal with ID migration}
\usage{
handle_id_migration(response)
}
\arguments{
\item{response}{The result of a GET() call to the Yelp API.}
}
\description{
If duplicate businesses are merged, one ID redirects to the other.
This manifests as an HTTP 301 code.
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/business_reviews#response-business-migrated}
}
