% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_yelp_business}
\alias{is_yelp_business}
\title{Is the input a Yelp business?}
\usage{
is_yelp_business(x)
}
\arguments{
\item{x}{An R variable.}
}
\value{
\code{TRUE} if the result is a data frame with a character column
named \code{business_id}.
}
\description{
Is the input a Yelp business?
}
