% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ISO_3166_2_CODES}
\alias{ISO_3166_2_CODES}
\title{ISO 3166-2 Codes}
\format{An object of class \code{character} of length 4835.}
\usage{
ISO_3166_2_CODES
}
\description{
All valid ISO 3166-2 state codes for each country.
}
\examples{
ISO_3166_2_CODES
}
\references{
Equal to \code{ISOcodes::ISO_3166_2$Code}. See
\code{\link[ISOcodes]{ISO_3166}} for details.
}
\keyword{datasets}
