% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone-search.R
\name{phone_search}
\alias{phone_search}
\title{Find businesses by phone number}
\usage{
phone_search(phone, access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{phone}{A string representing a phone number. This should include the
country code and consist of a plus sign followed by only digits, for example
"+1234567890".}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\value{
A data frame of business results. Note that there may be more than
one business returned, for example where multiple stores share a central
phone number.
}
\description{
Finds businesses with a particular phone number.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
call_me <- phone_search("+12127052000")
if(interactive()) View(call_me) else str(call_me)
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/business_search_phone}
}
