% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SUPPORTED_BUSINESS_ATTRIBUTES}
\alias{SUPPORTED_BUSINESS_ATTRIBUTES}
\title{Supported business attributes}
\format{An object of class \code{character} of length 6.}
\usage{
SUPPORTED_BUSINESS_ATTRIBUTES
}
\description{
The allowed values of business attribute, as used by the business search
\code{attribute} argument.
}
\examples{
SUPPORTED_BUSINESS_ATTRIBUTES
}
\references{
See the attributes field of the business search documentation.
\url{https://www.yelp.com/developers/documentation/v3/business_search}
}
\keyword{datasets}
