% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SUPPORTED_CATEGORY_ALIASES}
\alias{SUPPORTED_CATEGORY_ALIASES}
\title{Supported category aliases}
\format{An object of class \code{character} of length 1513.}
\usage{
SUPPORTED_CATEGORY_ALIASES
}
\description{
The allowed values of category alias, as used by the business search
\code{categories} argument.
}
\examples{
SUPPORTED_CATEGORY_ALIASES
}
\references{
Created via \code{data-raw/get-supported-category-aliases.R}
based on the values provided at
\url{https://www.yelp.com/developers/documentation/v3/all_category_list}
}
\keyword{datasets}
