% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SUPPORTED_LOCALES}
\alias{SUPPORTED_LOCALES}
\title{Supported category aliases}
\format{An object of class \code{character} of length 42.}
\usage{
SUPPORTED_LOCALES
}
\description{
The allowed values of locale, as used by the business search
\code{locale} argument.
}
\examples{
SUPPORTED_LOCALES
}
\references{
Created via \code{data-raw/get-supported-locales.R}
based on the values provided at
\url{https://www.yelp.com/developers/documentation/v3/supported_locales}
}
\seealso{
\code{\link{set_yelp_locale}}
}
\keyword{datasets}
