% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business-lookup.R
\name{business_lookup}
\alias{business_lookup}
\title{Look up a business by Yelp ID}
\usage{
business_lookup(business, locale = get_yelp_locale(),
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{locale}{A string naming the locale. See \code{\link{SUPPORTED_LOCALES}}
for allowed values.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}

\item{businesses}{A character vector describing the Yelp IDs of businesses,
as returned by \code{\link{business_search}},  or the data frame returned by
that API endpoint.}
}
\value{
A data frame with 28 columns. Each row corresponds to one business.
In addition to the fields provided by the business search, you also get
opening hours, URLs of up to 3 photos, whether or not the business has been
claimed by its owner, and whether or not the business has been permanently
closed.
}
\description{
Use the ID from a business search to get more details on that business.
}
\details{
If you pass multiple business IDs (either with a character vector
with length greater than one or a data frame with multiple rows), this will
result in multiple calls to the API, which you need to bear in mind if your
usage is limited.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
ghostbusters <- business_lookup("jDPKMwOtvPA0lkqRwRvSJQ")
if(interactive()) View(ghostbusters) else str(ghostbusters)
}
}
