% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{reviews}
\alias{reviews}
\title{Get reviews for a business}
\usage{
reviews(business_id, locale = "en_US",
  access_token = Sys.getenv("YELP_ACCESS_TOKEN", NA))
}
\arguments{
\item{business_id}{A string denoting a Yelp business ID, as returned by
the business search API.}

\item{locale}{A string naming the locale. See \code{\link{SUPPORTED_LOCALES}}
for allowed values.}

\item{access_token}{A string giving an access token to authenticate the API
call. See \code{\link{get_access_token}}.}
}
\value{
A data frame with 6 columns. Each row corresponds to one review of the
specified business.
}
\description{
Get the Yelp reviews associated with a business ID.
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
# First lookup businesses
theaters_in_chicago <- business_search("theater", "chicago")
# Get the reviews using the business ID
reviews_of_chicago_theater_id <- reviews(theaters_in_chicago$id[1L])
# ...or the alias
reviews_of_chicago_theater_alias <- reviews(theaters_in_chicago$alias[1L])
identical(
  reviews_of_chicago_theater_id,
  reviews_of_chicago_theater_alias
)
if(interactive())
  View(reviews_of_chicago_theater_id) else str(reviews_of_chicago_theater_id)
}
}
\references{
\url{https://www.yelp.com/developers/documentation/v3/business_reviews}
}
