% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{set_yelp_locale}
\alias{set_yelp_locale}
\title{Set a default locale}
\usage{
set_yelp_locale(locale)
}
\arguments{
\item{locale}{A string naming the locale. See \code{\link{SUPPORTED_LOCALES}}
for allowed values.}
}
\value{
Invoked for the side-effect of setting an environment variable to
store the default locale. The previous value of this variable is invisibly
returned.
}
\description{
Sets a default locale for the R session.
}
\examples{
set_yelp_locale("zh_HK")
}
